+++ query-engine/connectors/mongodb-query-connector/src/root_queries/mod.rs
@@ -1,6 +1,7 @@
 //! Top level queries to satisfy the connector interface operations.
 pub mod aggregate;
 pub mod read;
+mod update_utils;
 pub mod write;
 
 use crate::{output_meta::OutputMetaMapping, value::value_from_bson};
